avoid conversion in llvm ir

+++ commutative_loops/map_simple/map_simple.c
@@ -5,8 +5,8 @@
 int main() {
   unsigned array[] = {3, 5, 15, 99, 11, 33, 5, 67};
   const size_t N = sizeof(array) / sizeof(unsigned int);
-  unsigned j = 0;
-  unsigned k = 0;
+  size_t j = 0;
+  size_t k = 0;
 
   for (size_t i = 0; i < N; ++i) {
     array[i] = i;
